requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
mysql-connector-python==8.0.26
pyserial==3.5
python-dotenv==0.19.0
flask-migrate==3.1.0
flask-login==0.5.0
flask-wtf==1.0.0
wtforms==3.0.1